검색박스 구현

+++ src/main/webapp/WEB-INF/usr/article/list.jsp
@@ -1,5 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 
 <c:set var="pageTitle" value="게시물 리스트" />
@@ -10,14 +9,56 @@
 
 		<div class="card bordered shadow-lg">
 			<div class="card-title">
-				<a href="javascript:history.back();" class="cursor-pointer"> <i
-					class="fas fa-chevron-left"></i>
-				</a> <span>게시물 리스트</span>
+				<a href="javascript:history.back();" class="cursor-pointer">
+					<i class="fas fa-chevron-left"></i>
+				</a>
+				<span>게시물 리스트</span>
 			</div>
 
 			<div class="px-4 py-4">
-				<div class="badge badge-primary">전체게시물 개수</div>
+				<div class="badge badge-primary">
+					<c:if test="${param.searchKeyword == null}">
+						전체게시물 개수
+					</c:if>
+					<c:if test="${param.searchKeyword != null}">
+						검색어 `${param.searchKeyword}`, 게시물 개수
+					</c:if>
+				</div>
 				${totalItemsCount}
+			</div>
+
+			<hr />
+
+			<div class="px-4 py-4">
+				<div class="badge badge-primary">검색</div>
+				<form action="">
+					<input type="hidden" name="boardId" value="${boardId}" />
+					<div class="form-control">
+						<label class="label">
+							<span class="label-text">검색타입</span>
+						</label>
+						<div>
+							<select name="searchKeywordTypeCode" class="select select-bordered w-full max-w-md">
+								<option value="title,body">제목,내용</option>
+								<option value="title">제목</option>
+								<option value="title">내용</option>
+							</select>
+						</div>
+					</div>
+
+					<div class="form-control">
+						<label class="label">
+							<span class="label-text">검색어</span>
+						</label>
+						<div>
+							<input class="input input-bordered w-full max-w-md" maxlength="100" name="searchKeyword" type="text" placeholder="검색어를 입력해주세요." value="${param.searchKeyword}" />
+						</div>
+					</div>
+
+					<div class="btns">
+						<button type="submit" class="btn btn-link">검색</button>
+					</div>
+				</form>
 			</div>
 
 			<hr />
@@ -28,33 +69,39 @@
 
 					<div class="py-4">
 						<div class="grid gap-3" style="grid-template-columns: 100px 1fr;">
-							<a href="${detailUri}"> <img class="rounded-full w-full"
-								src="https://i.pravatar.cc/200?img=37" alt="">
-							</a> <a href="${detailUri}" class="hover:underline cursor-pointer">
+							<a href="${detailUri}">
+								<img class="rounded-full w-full" src="https://i.pravatar.cc/200?img=37" alt="">
+							</a>
+							<a href="${detailUri}" class="hover:underline cursor-pointer">
 								<span class="badge badge-outline">제목</span>
 								<div class="line-clamp-3">${article.titleForPrint}</div>
 							</a>
 						</div>
 
 						<div class="mt-3 grid sm:grid-cols-2 lg:grid-cols-4 gap-3">
-							<a href="${detailUri}" class="hover:underline"> <span
-								class="badge badge-primary">번호</span> <span>${article.id}</span>
-							</a> <a href="${detailUri}" class="cursor-pointer hover:underline">
-								<span class="badge badge-accent">작성자</span> <span>${article.extra__writerName}</span>
-							</a> <a href="${detailUri}" class="hover:underline"> <span
-								class="badge">등록날짜</span> <span class="text-gray-600 text-light">${article.regDate}</span>
-							</a> <a href="${detailUri}" class="hover:underline"> <span
-								class="badge">수정날짜</span> <span class="text-gray-600 text-light">${article.updateDate}</span>
+							<a href="${detailUri}" class="hover:underline">
+								<span class="badge badge-primary">번호</span>
+								<span>${article.id}</span>
+							</a>
+							<a href="${detailUri}" class="cursor-pointer hover:underline">
+								<span class="badge badge-accent">작성자</span>
+								<span>${article.extra__writerName}</span>
+							</a>
+							<a href="${detailUri}" class="hover:underline">
+								<span class="badge">등록날짜</span>
+								<span class="text-gray-600 text-light">${article.regDate}</span>
+							</a>
+							<a href="${detailUri}" class="hover:underline">
+								<span class="badge">수정날짜</span>
+								<span class="text-gray-600 text-light">${article.updateDate}</span>
 							</a>
 						</div>
 
-						<a href="${detailUri}"
-							class="block mt-3 hover:underline cursor-pointer col-span-1 sm:col-span-2 xl:col-span-3">
+						<a href="${detailUri}" class="block mt-3 hover:underline cursor-pointer col-span-1 sm:col-span-2 xl:col-span-3">
 							<span class="badge badge-outline">본문</span>
 
 							<div class="mt-2">
-								<img class="rounded" src="https://picsum.photos/id/237/300/300"
-									alt="" />
+								<img class="rounded" src="https://picsum.photos/id/237/300/300" alt="" />
 							</div>
 
 							<div class="line-clamp-3">${article.bodySummaryForPrint}</div>
@@ -64,13 +111,18 @@
 					<div class="btns mt-3">
 						<c:if test="${article.extra__actorCanModify}">
 							<a href="../article/modify?id=${article.id}" class="btn btn-link">
-								<span><i class="fas fa-edit"></i></span> <span>수정</span>
+								<span>
+									<i class="fas fa-edit"></i>
+								</span>
+								<span>수정</span>
 							</a>
 						</c:if>
 						<c:if test="${article.extra__actorCanDelete}">
-							<a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;"
-								href="../article/doDelete?id=${article.id}" class="btn btn-link">
-								<span><i class="fas fa-trash-alt"></i></span> <span>삭제</span>
+							<a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" href="../article/doDelete?id=${article.id}" class="btn btn-link">
+								<span>
+									<i class="fas fa-trash-alt"></i>
+								</span>
+								<span>삭제</span>
 							</a>
 						</c:if>
 					</div>
@@ -80,12 +132,10 @@
 
 				<div class="page-menu">
 					<c:set var="baseUri" value="?boardId=${boardId}" />
-				
+
 					<c:set var="pageMenuArmSize" value="7" />
-					<c:set var="startPage"
-						value="${page - pageMenuArmSize >= 1 ? page - pageMenuArmSize : 1}" />
-					<c:set var="endPage"
-						value="${page + pageMenuArmSize <= totalPage ? page + pageMenuArmSize : totalPage}" />
+					<c:set var="startPage" value="${page - pageMenuArmSize >= 1 ? page - pageMenuArmSize : 1}" />
+					<c:set var="endPage" value="${page + pageMenuArmSize <= totalPage ? page + pageMenuArmSize : totalPage}" />
 
 					<div class="btn-group">
 						<c:if test="${startPage > 1}">
